+++ CHANGELOG.md
@@ -30,6 +30,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support Mainnet for Lido CSM setup.
 
+### Changed
+- JWT secret path set as relative path when not provided.
+
 ## [v1.6.0] - 2024-10-18